move some hooks to hook/utils.js
remove hook/fade
